Move SingleGamepadAxis and DualGamepadAxis to axislike.rs
